Create KeePassXC sync application
Add installation script for Ubuntu and install 2 new commands:
keepassxc-sync-cli and keepassxc-sync.

keepassxc-sync-cli just run the original script. Note, this script
was renamed from : keepassxc-sync.sh to keepassxc-sync-cli.bash.

keepassxc-sync is an Ubuntu application that start Keepass XC and
synchronize the databases at the startup and end of Keepass XC.